+++ src/main/java/com/minecraftonline/griefalert/storage/SqlProfileStorage.java
@@ -37,10 +37,11 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Optional;
 import javax.annotation.Nonnull;
+
+import org.jetbrains.annotations.Nullable;
 import org.spongepowered.api.Sponge;
 import org.spongepowered.api.service.sql.SqlService;
 import org.spongepowered.api.text.format.TextColor;
-import org.spongepowered.api.world.DimensionTypes;
 
 /**
  * Deprecated ProfileStorage implementation.
@@ -126,6 +127,13 @@ public class SqlProfileStorage implements ProfileStorage {
       statement.execute();
     }
     return true;
+  }
+
+  @Nullable
+  @Override
+  public GriefProfile get(@Nonnull GriefEvent griefEvent, @Nonnull String target) throws Exception {
+    // TODO implement
+    return null;
   }
 
   @Nonnull